ws2 lab

+++ workshops/ws2/lab/main/shopping.c
@@ -12,25 +12,25 @@ void shoppingStats(int noOfItems) {
 // noOfItem: the number of items purchased
     int loop;
     //integer for number of loops
-    float price;
+    double price;
     //float value for item prices
-    float sum;
+    double sum;
     //float for total price of items
-    float avg;
+    double avg;
     //float for average price of items
     printf("Please enter the price of the %d items bought...\n",noOfItems);
     loop = 0;
     sum = 0;
     while (loop != noOfItems) {
         printf("%d> ", loop+1);
-        scanf("%f", &price);
+        scanf("%lf", &price);
         loop++;
         sum += price;
         }
     line('-',50);
-    printf("Total price for all the items purchased: %.2f\n", sum);
+    printf("Total price for all the items purchased: %.2lf\n", sum);
     avg = sum / noOfItems;
-    printf("Average price of the items: %.2f\n",avg);
+    printf("Average price of the items: %.2lf\n",avg);
     line('=',50);
 }
